PLUMBERS: Remove need for highres
Downscale images to 320x240 if highres is not supported.
