+++ engines/plumbers/POTFILES
@@ -1,0 +1,1 @@
+engines/plumbers/metaengine.cpp
