+++ engines/plumbers/module.mk
@@ -3,7 +3,12 @@ MODULE := engines/plumbers
 MODULE_OBJS = \
 	plumbers.o \
 	console.o \
-	metaengine.o
+	metaengine.o \
+	windows.o
+
+ifdef USE_RGB_COLOR
+MODULE_OBJS += 3do.o
+endif
 
 # This module can be built as a plugin
 ifeq ($(ENABLE_PLUMBERS), DYNAMIC_PLUGIN)
